+++ examples/demo-viewer/agent/demo/crewai_agentic_chat/page.tsx
@@ -2,9 +2,9 @@
 import React, { useState } from "react";
 import "@copilotkit/react-ui/styles.css";
 import "./style.css";
-import { CopilotKit, useCopilotAction } from "@copilotkit/react-core";
-import { CopilotChat } from "@copilotkit/react-ui";
-
+import { CopilotKit, useCopilotAction, useCopilotChat } from "@copilotkit/react-core";
+import { CopilotChat, useCopilotChatSuggestions } from "@copilotkit/react-ui";
+import { chatSuggestions, initialPrompt } from "@/lib/prompts";
 const AgenticChat: React.FC = () => {
   return (
     <CopilotKit
@@ -37,15 +37,21 @@ const Chat = () => {
     },
   });
 
+  useCopilotChatSuggestions({
+    instructions: chatSuggestions.agenticChat,
+    // className : "bg-gray-100"
+  })
+
+
   return (
     <div
       className="flex justify-center items-center h-full w-full"
       style={{ background }}
     >
-      <div className="w-8/10 h-8/10 rounded-lg">
+      <div className="w-8/10 h-8/10 rounded-lg ">
         <CopilotChat
-          className="h-full rounded-2xl"
-          labels={{ initial: "Hi, I'm an agent. Want to chat?" }}
+          className="h-full w-full rounded-2xl py-6"
+          labels={{ initial: initialPrompt.agenticChat }}
         />
       </div>
     </div>